Reorganize/address comments/add team and MDRange to changes
New structure: TetCommonPolicyConstructors for things common to all
policies, Test[Range,Team,MDRange]PolicyConstructors for thing specific
to a policy
