+++ core/unit_test/TestInit.hpp
@@ -42,6 +42,3 @@ TEST(TEST_CATEGORY, dispatch) { test_dispatch<TEST_EXECSPACE>(); }
 }  // namespace Test
 
 #include <TestCompilerMacros.hpp>
-#ifndef KOKKOS_ENABLE_OPENACC  // FIXME_OPENACC TeamPolicy is not implemented
-#include <TestPolicyConstruction.hpp>
-#endif
